# packages/sdk/openapi.json
{
  "components": {
    "schemas": {
      "APIError": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/APIError.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "ConsumerPreferencesData": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ConsumerPreferencesData.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "consumerId": {
            "type": "string"
          },
          "favoriteProfessionalIds": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "resolvedLocation": {
            "$ref": "#/components/schemas/ResolvedLocationData"
          },
          "selectedAreaId": {
            "type": "string"
          },
          "userLocation": {
            "$ref": "#/components/schemas/GeoPointData"
          }
        },
        "required": [
          "favoriteProfessionalIds",
          "resolvedLocation",
          "userLocation"
        ],
        "type": "object"
      },
      "ConsumerPreferencesResponseBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ConsumerPreferencesResponseBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ConsumerPreferencesData"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "CustomerAuthCreateSessionRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/CustomerAuthCreateSessionRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "phone",
          "password"
        ],
        "type": "object"
      },
      "CustomerAuthRegisterRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/CustomerAuthRegisterRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "displayName",
          "password",
          "phone"
        ],
        "type": "object"
      },
      "CustomerAuthSessionData": {
        "additionalProperties": false,
        "properties": {
          "city": {
            "type": "string"
          },
          "consumerId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          },
          "isAuthenticated": {
            "type": "boolean"
          },
          "lastLoginAt": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "registeredAt": {
            "type": "string"
          }
        },
        "required": [
          "consumerId",
          "isAuthenticated",
          "phone"
        ],
        "type": "object"
      },
      "CustomerAuthSessionResponseBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/CustomerAuthSessionResponseBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/CustomerAuthSessionData"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "CustomerNotificationStateData": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/CustomerNotificationStateData.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "readIds": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "readIds"
        ],
        "type": "object"
      },
      "CustomerNotificationsResponseBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/CustomerNotificationsResponseBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/CustomerNotificationStateData"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "ErrorPayload": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "GeoPointData": {
        "additionalProperties": false,
        "properties": {
          "latitude": {
            "format": "double",
            "type": "number"
          },
          "longitude": {
            "format": "double",
            "type": "number"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ],
        "type": "object"
      },
...
      "ResolvedLocationData": {
        "additionalProperties": false,
        "properties": {
          "areaId": {
            "type": "string"
          },
          "areaLabel": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "formattedAddress": {
            "type": "string"
          },
          "point": {
            "$ref": "#/components/schemas/GeoPointData"
          },
          "postalCode": {
            "type": "string"
          },
          "precision": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "source": {
            "type": "string"
          }
        },
        "required": [
          "areaId",
          "areaLabel",
          "city",
          "country",
          "district",
          "formattedAddress",
          "point",
          "postalCode",
          "precision",
          "province",
          "source"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AdminSessionAuth": {
        "description": "Admin session cookie issued by POST /api/v1/admin/auth/session.",
        "in": "cookie",
        "name": "bidanapp_admin_session",
        "type": "apiKey"
      },
      "CustomerSessionAuth": {
        "description": "Customer session cookie issued by POST /api/v1/customers/auth/session.",
        "in": "cookie",
        "name": "bidanapp_customer_session",
        "type": "apiKey"
      },
      "ProfessionalSessionAuth": {
        "description": "Professional session cookie issued by POST /api/v1/professionals/auth/session.",
        "in": "cookie",
        "name": "bidanapp_professional_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Backend-generated OpenAPI contract for BidanApp.",
    "title": "bidanapp-backend",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/consumers/preferences": {
      "get": {
        "operationId": "get-consumer-preferences",
        "parameters": [
          {
            "description": "Optional consumer id to scope consumer preference state",
            "explode": false,
            "in": "query",
            "name": "consumer_id",
            "schema": {
              "description": "Optional consumer id to scope consumer preference state",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumerPreferencesResponseBody"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "CustomerSessionAuth": []
          }
        ],
        "summary": "Get persisted consumer preference state",
        "tags": [
          "App State"
        ]
      },
      "put": {
        "operationId": "upsert-consumer-preferences",
        "parameters": [
          {
            "description": "Optional consumer id to scope consumer preference state",
            "explode": false,
            "in": "query",
            "name": "consumer_id",
            "schema": {
              "description": "Optional consumer id to scope consumer preference state",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumerPreferencesData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumerPreferencesResponseBody"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "CustomerSessionAuth": []
          }
        ],
        "summary": "Persist consumer preference state",
        "tags": [
          "App State"
        ]
      }
    },
...
    "/customers/auth/register": {
      "post": {
        "operationId": "register-customer-auth-account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerAuthRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerAuthSessionResponseBody"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Bad Request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Register a customer account",
        "tags": [
          "Customer Auth"
        ]
      }
    },
    "/customers/auth/session": {
      "delete": {
        "operationId": "delete-customer-auth-session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerAuthSessionResponseBody"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "CustomerSessionAuth": []
          }
        ],
        "summary": "Revoke the authenticated customer session",
        "tags": [
          "Customer Auth"
        ]
      },
      "get": {
        "operationId": "get-customer-auth-session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerAuthSessionResponseBody"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "CustomerSessionAuth": []
          }
        ],
        "summary": "Get the authenticated customer session",
        "tags": [
          "Customer Auth"
        ]
      },
      "post": {
        "operationId": "create-customer-auth-session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerAuthCreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerAuthSessionResponseBody"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create an authenticated customer session",
        "tags": [
          "Customer Auth"
        ]
      }
    },
...
    "/notifications/customer": {
      "get": {
        "operationId": "get-customer-notifications-state",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerNotificationsResponseBody"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "CustomerSessionAuth": []
          }
        ],
        "summary": "Get persisted customer notification read state",
        "tags": [
          "Notifications"
        ]
      },
      "put": {
        "operationId": "upsert-customer-notifications-state",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerNotificationStateData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerNotificationsResponseBody"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "CustomerSessionAuth": []
          }
        ],
        "summary": "Persist customer notification read state",
        "tags": [
          "Notifications"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}